bybit-bot/requirements.txt
pybit>=5.6.0
python-dotenv>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
